DeepCovVar_Web/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
biopython==1.81
pandas==1.5.3
numpy==1.24.4
openpyxl==3.1.2
fpdf2==2.7.6
requests==2.31.0
Werkzeug==2.3.7
